refactor(DC-2174): introduce prometheus url as input variable
